pyproject.toml
[project]
name = "mock-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.11",
]
